[WFGP-182] Use Batavia facade API to transform EE9 artifacts instead of Eclipse transformer directly.
