+++ transformer/pom.xml
@@ -34,10 +34,14 @@
 
   <dependencies>
     <dependency>
-      <groupId>org.eclipse.transformer</groupId>
-      <artifactId>org.eclipse.transformer</artifactId>
+      <groupId>org.wildfly.extras.batavia</groupId>
+      <artifactId>transformer-api</artifactId>
     </dependency>
-
+    <dependency>
+      <groupId>org.wildfly.extras.batavia</groupId>
+      <artifactId>transformer-impl-eclipse</artifactId>
+      <scope>runtime</scope>
+    </dependency>
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
