--- transformer/src/main/resources/org/wildfly/galleon/plugin/transformer/cdi-annotations.properties
@@ -1,1 +1,0 @@
-javax=jakarta
