--- transformer/src/main/resources/org/wildfly/galleon/plugin/transformer/jakarta-direct.properties
@@ -1,2 +1,0 @@
-javax.jms.Queue=jakarta.jms.Queue
-javax.jms.Topic=jakarta.jms.Topic
