--- transformer/src/main/resources/org/wildfly/galleon/plugin/transformer/jakarta-per-class.properties
@@ -1,3 +1,0 @@
-org/jboss/as/jaxrs/JaxrsAnnotations$Constants.class=jaxrs-annotations.properties
-org/apache/jasper/compiler/Generator.class=jsp-compiler.properties
-org/jboss/as/weld/CdiAnnotations$Constants.class=cdi-annotations.properties
