--- transformer/src/main/resources/org/wildfly/galleon/plugin/transformer/jakarta-renames.properties
@@ -1,174 +1,0 @@
-# Default jakarta configuration for package renames.
-#
-# Last update: 11-Jun-2020
-#
-# Format:
-# currentPackageName=newPackageName
-#
-# The current package name may contain a ".*" suffix.  The suffix indicates
-# that sub-packages are to be matched.  For example:
-#
-# javax.servlet.*=jakartax.servlet
-#
-# This will cause sub-packages like javax.servlet.annotation
-# to be replaced with jakartax.servlet.annotation
-#
-# By default, sub-packages are not updated.
-#
-# Package renames apply in several ways:
-#   Class references in .class files are updated.
-#   Qualified class names which appear in java string constants in .class files are updated.
-#   Qualified resource references which appear in java string constants in .class files are updated.
-#   Package references in manifest (MANIFEST.MF) and feature manifest (featureName.mf) files are updated.
-#
-
-javax.activation=jakarta.activation
-
-javax.annotation=jakarta.annotation
-javax.annotation.security=jakarta.annotation.security
-javax.annotation.sql=jakarta.annotation.sql
-javax.batch=jakarta.batch
-javax.batch.api=jakarta.batch.api
-javax.batch.api.chunk=jakarta.batch.api.chunk
-javax.batch.api.chunk.listener=jakarta.batch.api.chunk.listener
-javax.batch.api.listener=jakarta.batch.api.listener
-javax.batch.api.partition=jakarta.batch.api.partition
-javax.batch.operations=jakarta.batch.operations
-javax.batch.runtime=jakarta.batch.runtime
-javax.batch.runtime.context=jakarta.batch.runtime.context
-
-com.ibm.wsspi.el=com.ibm.wsspi.el
-
-javax.decorator=jakarta.decorator
-javax.ejb=jakarta.ejb
-javax.ejb.embeddable=jakarta.ejb.embeddable
-javax.ejb.spi=jakarta.ejb.spi
-javax.el=jakarta.el
-
-javax.enterprise.concurrent=jakarta.enterprise.concurrent
-javax.enterprise.context=jakarta.enterprise.context
-javax.enterprise.context.control=jakarta.enterprise.context.control
-javax.enterprise.context.spi=jakarta.enterprise.context.spi
-javax.enterprise.event=jakarta.enterprise.event
-javax.enterprise.inject=jakarta.enterprise.inject
-javax.enterprise.inject.literal=jakarta.enterprise.inject.literal
-javax.enterprise.inject.se=jakarta.enterprise.inject.se
-javax.enterprise.inject.spi=jakarta.enterprise.inject.spi
-javax.enterprise.inject.spi.configurator=jakarta.enterprise.inject.spi.configurator
-javax.enterprise.util=jakarta.enterprise.util
-javax.faces=jakarta.faces
-javax.faces.annotation=jakarta.faces.annotation
-javax.faces.application=jakarta.faces.application
-javax.faces.bean=jakarta.faces.bean
-javax.faces.component=jakarta.faces.component
-javax.faces.component.behavior=jakarta.faces.component.behavior
-javax.faces.component.html=jakarta.faces.component.html
-javax.faces.component.search=jakarta.faces.component.search
-javax.faces.component.visit=jakarta.faces.component.visit
-javax.faces.context=jakarta.faces.context
-javax.faces.convert=jakarta.faces.convert
-javax.faces.el=jakarta.faces.el
-javax.faces.event=jakarta.faces.event
-javax.faces.flow=jakarta.faces.flow
-javax.faces.flow.builder=jakarta.faces.flow.builder
-javax.faces.lifecycle=jakarta.faces.lifecycle
-javax.faces.model=jakarta.faces.model
-javax.faces.push=jakarta.faces.push
-javax.faces.render=jakarta.faces.render
-javax.faces.validator=jakarta.faces.validator
-javax.faces.view=jakarta.faces.view
-javax.faces.view.facelets=jakarta.faces.view.facelets
-javax.faces.webapp=jakarta.faces.webapp
-javax.inject=jakarta.inject
-
-javax.interceptor=jakarta.interceptor
-javax.jms=jakarta.jms
-javax.json=jakarta.json
-javax.json.bind=jakarta.json.bind
-javax.json.bind.adapter=jakarta.json.bind.adapter
-javax.json.bind.annotation=jakarta.json.bind.annotation
-javax.json.bind.config=jakarta.json.bind.config
-javax.json.bind.serializer=jakarta.json.bind.serializer
-javax.json.bind.spi=jakarta.json.bind.spi
-javax.json.spi=jakarta.json.spi
-javax.json.stream=jakarta.json.stream
-javax.jws=jakarta.jws
-javax.jws.soap=jakarta.jws.soap
-javax.mail=jakarta.mail
-javax.mail.event=jakarta.mail.event
-javax.mail.internet=jakarta.mail.internet
-javax.mail.search=jakarta.mail.search
-javax.mail.util=jakarta.mail.util
-javax.persistence=jakarta.persistence
-javax.persistence.criteria=jakarta.persistence.criteria
-javax.persistence.metamodel=jakarta.persistence.metamodel
-javax.persistence.spi=jakarta.persistence.spi
-
-javax.resource=jakarta.resource
-javax.resource.cci=jakarta.resource.cci
-javax.resource.spi=jakarta.resource.spi
-javax.resource.spi.endpoint=jakarta.resource.spi.endpoint
-javax.resource.spi.security=jakarta.resource.spi.security
-javax.resource.spi.work=jakarta.resource.spi.work
-javax.security.auth.message=jakarta.security.auth.message
-javax.security.auth.message.callback=jakarta.security.auth.message.callback
-javax.security.auth.message.config=jakarta.security.auth.message.config
-javax.security.auth.message.module=jakarta.security.auth.message.module
-javax.security.enterprise=jakarta.security.enterprise
-javax.security.enterprise.authentication=jakarta.security.enterprise.authentication
-javax.security.enterprise.authentication.mechanism=jakarta.security.enterprise.authentication.mechanism
-javax.security.enterprise.authentication.mechanism.http=jakarta.security.enterprise.authentication.mechanism.http
-javax.security.enterprise.credential=jakarta.security.enterprise.credential
-javax.security.enterprise.identitystore=jakarta.security.enterprise.identitystore
-javax.security.jacc=jakarta.security.jacc
-javax.servlet=jakarta.servlet
-javax.servlet.annotation=jakarta.servlet.annotation
-javax.servlet.descriptor=jakarta.servlet.descriptor
-javax.servlet.http=jakarta.servlet.http
-javax.servlet.jsp=jakarta.servlet.jsp
-javax.servlet.jsp.el=jakarta.servlet.jsp.el
-javax.servlet.jsp.jstl=jakarta.servlet.jsp.jstl
-javax.servlet.jsp.jstl.core=jakarta.servlet.jsp.jstl.core
-javax.servlet.jsp.jstl.fmt=jakarta.servlet.jsp.jstl.fmt
-javax.servlet.jsp.jstl.sql=jakarta.servlet.jsp.jstl.sql
-javax.servlet.jsp.jstl.tlv=jakarta.servlet.jsp.jstl.tlv
-javax.servlet.jsp.resources=jakarta.servlet.jsp.resources
-javax.servlet.jsp.tagext=jakarta.servlet.jsp.tagext
-javax.servlet.resources=jakarta.servlet.resources
-javax.servlet.jsp.tagext.JspFragment=jakarta.servlet.jsp.tagext.JspFragment
-
-
-javax.transaction=jakarta.transaction
-javax.validation=jakarta.validation
-javax.validation.bootstrap=jakarta.validation.bootstrap
-javax.validation.constraints=jakarta.validation.constraints
-javax.validation.constraintvalidation=jakarta.validation.constraintvalidation
-javax.validation.executable=jakarta.validation.executable
-javax.validation.groups=jakarta.validation.groups
-javax.validation.metadata=jakarta.validation.metadata
-javax.validation.spi=jakarta.validation.spi
-javax.validation.valueextraction=jakarta.validation.valueextraction
-javax.websocket=jakarta.websocket
-javax.websocket.server=jakarta.websocket.server
-javax.ws.rs=jakarta.ws.rs
-javax.ws.rs.client=jakarta.ws.rs.client
-javax.ws.rs.container=jakarta.ws.rs.container
-javax.ws.rs.core=jakarta.ws.rs.core
-javax.ws.rs.ext=jakarta.ws.rs.ext
-javax.ws.rs.sse=jakarta.ws.rs.sse
-javax.xml.bind=jakarta.xml.bind
-javax.xml.bind.annotation=jakarta.xml.bind.annotation
-javax.xml.bind.annotation.adapters=jakarta.xml.bind.annotation.adapters
-javax.xml.bind.attachment=jakarta.xml.bind.attachment
-javax.xml.bind.helpers=jakarta.xml.bind.helpers
-javax.xml.bind.util=jakarta.xml.bind.util
-javax.xml.soap=jakarta.xml.soap
-javax.xml.ws=jakarta.xml.ws
-javax.xml.ws.handler=jakarta.xml.ws.handler
-javax.xml.ws.handler.soap=jakarta.xml.ws.handler.soap
-javax.xml.ws.http=jakarta.xml.ws.http
-javax.xml.ws.soap=jakarta.xml.ws.soap
-javax.xml.ws.spi=jakarta.xml.ws.spi
-javax.xml.ws.spi.http=jakarta.xml.ws.spi.http
-javax.xml.ws.wsaddressing=jakarta.xml.ws.wsaddressing
-
