--- transformer/src/main/resources/org/wildfly/galleon/plugin/transformer/jakarta-txt-master.properties
@@ -1,6 +1,0 @@
-*.xml=jakarta-renames.properties
-*.tld=jakarta-renames.properties
-pom.xml=pom-xml.properties
-ejb-jar.xml=jakarta-renames.properties
-jboss-ejb3.xml=jakarta-renames.properties
-ra.xml=jakarta-renames.properties
