--- transformer/src/main/resources/org/wildfly/galleon/plugin/transformer/jaxrs-annotations.properties
@@ -1,1 +1,0 @@
-javax=jakarta
