--- transformer/src/main/resources/org/wildfly/galleon/plugin/transformer/jsp-compiler.properties
@@ -1,1 +1,0 @@
-javax.servlet=jakarta.servlet
