--- transformer/src/main/resources/org/wildfly/galleon/plugin/transformer/pom-xml.properties
@@ -1,1 +1,0 @@
-
